+++ tfe/resource_tfe_workspace_test.go
@@ -8,88 +8,9 @@ import (
 
 	tfe "github.com/hashicorp/go-tfe"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/terraform"
 )
-
-func TestPackWorkspaceID(t *testing.T) {
-	cases := []struct {
-		w   *tfe.Workspace
-		id  string
-		err bool
-	}{
-		{
-			w: &tfe.Workspace{
-				Name: "my-workspace-name",
-				Organization: &tfe.Organization{
-					Name: "my-org-name",
-				},
-			},
-			id:  "my-org-name/my-workspace-name",
-			err: false,
-		},
-		{
-			w: &tfe.Workspace{
-				Name: "my-workspace-name",
-			},
-			id:  "",
-			err: true,
-		},
-	}
-
-	for _, tc := range cases {
-		id, err := packWorkspaceID(tc.w)
-		if (err != nil) != tc.err {
-			t.Fatalf("expected error is %t, got %v", tc.err, err)
-		}
-
-		if tc.id != id {
-			t.Fatalf("expected ID %q, got %q", tc.id, id)
-		}
-	}
-}
-
-func TestUnpackWorkspaceID(t *testing.T) {
-	cases := []struct {
-		id   string
-		org  string
-		name string
-		err  bool
-	}{
-		{
-			id:   "my-org-name/my-workspace-name",
-			org:  "my-org-name",
-			name: "my-workspace-name",
-			err:  false,
-		},
-		{
-			id:   "my-workspace-name|my-org-name",
-			org:  "my-org-name",
-			name: "my-workspace-name",
-			err:  false,
-		},
-		{
-			id:   "some-invalid-id",
-			org:  "",
-			name: "",
-			err:  true,
-		},
-	}
-
-	for _, tc := range cases {
-		org, name, err := unpackWorkspaceID(tc.id)
-		if (err != nil) != tc.err {
-			t.Fatalf("expected error is %t, got %v", tc.err, err)
-		}
-
-		if tc.org != org {
-			t.Fatalf("expected organization %q, got %q", tc.org, org)
-		}
-
-		if tc.name != name {
-			t.Fatalf("expected name %q, got %q", tc.name, name)
-		}
-	}
-}
 
 func TestAccTFEWorkspace_basic(t *testing.T) {
 	workspace := &tfe.Workspace{}
@@ -119,6 +40,25 @@ func TestAccTFEWorkspace_basic(t *testing.T) {
 						"tfe_workspace.foobar", "trigger_prefixes.#", "0"),
 					resource.TestCheckResourceAttr(
 						"tfe_workspace.foobar", "working_directory", ""),
+				),
+			},
+		},
+	})
+}
+
+func TestAccTFEWorkspace_panic(t *testing.T) {
+	resource.Test(t, resource.TestCase{
+		PreCheck:     func() { testAccPreCheck(t) },
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckTFEWorkspaceDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config:             testAccTFEWorkspace_basic,
+				ExpectNonEmptyPlan: true,
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckTFEWorkspaceExists(
+						"tfe_workspace.foobar", &tfe.Workspace{}),
+					testAccCheckTFEWorkspacePanic("tfe_workspace.foobar"),
 				),
 			},
 		},
@@ -462,27 +402,50 @@ func testAccCheckTFEWorkspaceExists(
 			return fmt.Errorf("No instance ID is set")
 		}
 
-		// Get the organization and workspace name.
-		organization, name, err := unpackWorkspaceID(rs.Primary.ID)
-		if err != nil {
-			return fmt.Errorf("Error unpacking workspace ID: %v", err)
-		}
-
-		w, err := tfeClient.Workspaces.Read(ctx, organization, name)
+		// Get the workspace
+		w, err := tfeClient.Workspaces.ReadByID(ctx, rs.Primary.ID)
 		if err != nil {
 			return err
 		}
 
-		id, err := packWorkspaceID(w)
-		if err != nil {
-			return fmt.Errorf("Error creating ID for workspace %s: %v", name, err)
-		}
-
-		if id != rs.Primary.ID {
-			return fmt.Errorf("Workspace not found")
-		}
-
 		*workspace = *w
+
+		return nil
+	}
+}
+
+// As of 4/20/2020 there is a bug that will cause the provider to panic
+// if a workspace is deleted outside of terraform. This case is handled
+// by the data_workspace but not resource_workspace.
+//
+// This test demonstrates the bug.
+//
+// panic: runtime error: invalid memory address or nil pointer dereference
+// resource_tfe_workspace.go:208 resourceTFEWorkspaceRead(...)
+func testAccCheckTFEWorkspacePanic(n string) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		tfeClient := testAccProvider.Meta().(*tfe.Client)
+
+		// Grab the resource out of the state and delete it from TFC/E directly.
+		rs, ok := s.RootModule().Resources[n]
+		if !ok {
+			return fmt.Errorf("Not found: %s", n)
+		}
+
+		err := tfeClient.Workspaces.DeleteByID(ctx, rs.Primary.ID)
+		if err != nil {
+			return fmt.Errorf("Could not delete %s: %v", n, err)
+		}
+
+		// Read the workspace again using the lower level resource reader
+		// which will trigger the panic
+		rd := &schema.ResourceData{}
+		rd.SetId(rs.Primary.ID)
+
+		err = resourceTFEWorkspaceRead(rd, testAccProvider.Meta())
+		if err != nil && err != tfe.ErrResourceNotFound {
+			return fmt.Errorf("Could not re-read resource directly: %v", err)
+		}
 
 		return nil
 	}
@@ -654,13 +617,7 @@ func testAccCheckTFEWorkspaceDestroy(s *terraform.State) error {
 			return fmt.Errorf("No instance ID is set")
 		}
 
-		// Get the organization and workspace name.
-		organization, name, err := unpackWorkspaceID(rs.Primary.ID)
-		if err != nil {
-			return fmt.Errorf("Error unpacking workspace ID: %v", err)
-		}
-
-		_, err = tfeClient.Workspaces.Read(ctx, organization, name)
+		_, err := tfeClient.Workspaces.ReadByID(ctx, rs.Primary.ID)
 		if err == nil {
 			return fmt.Errorf("Workspace %s still exists", rs.Primary.ID)
 		}